+++ lib/src/util/dart_files.dart
@@ -1,7 +1,10 @@
 import 'dart:io';
 
 import 'package:dart_dependency_checker/src/dependency_type.dart';
+import 'package:dart_dependency_checker/src/util/iterable_ext.dart';
 
+const _dartFileExt = '.dart';
+const _importPackagePattern = 'import \'package:';
 final _importPackageExp = RegExp(r':(.*?)/');
 
 abstract final class DartFiles {
@@ -13,19 +16,19 @@ abstract final class DartFiles {
 
   static Set<String> packages(File file) => file
       .readAsLinesSync()
-      .where((line) => line.startsWith('import \'package:'))
+      .where((line) => line.startsWith(_importPackagePattern))
       .map((import) => _importPackageExp.firstMatch(import)?[1])
       .nonNulls
-      .toSet();
+      .unmodifiable;
 
   static Set<File> _fromPath(String path) {
     final directory = Directory(path);
     if (directory.existsSync()) {
       return directory
           .listSync(recursive: true, followLinks: false)
-          .where((file) => file is File && file.path.endsWith('.dart'))
+          .where((file) => file is File && file.path.endsWith(_dartFileExt))
           .map((file) => file as File)
-          .toSet();
+          .unmodifiable;
     }
     return const {};
   }