+++ lib/src/util/iterable_ext.dart
@@ -1,0 +1,3 @@
+extension IterableExt<T> on Iterable<T> {
+  Set<T> get unmodifiable => Set.unmodifiable(this);
+}