+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 0.1.7
 
-- Update docs, dependencies and internals.
 - `TransitiveUseChecker` ignores:
     - own referenced package.
     - provided in main but missing in dev dependencies.
+- Update docs, dependencies and internals.
 - TODO(albert): remove before release.
 
 ## 0.1.6